Update recreate.py, proposing pwinput in place of getpass
pwinput provide asterisks in place of letters instead of masking all visual feedback using getpass. However, this maybe less secure for password because it reveals its length. But if we want to have a password minimum length as restriction, pwinput will help user to count the entered letter for the password.